requirements.txt
Flask==2.3.3
google-cloud-storage==2.10.0
google-cloud-bigquery==3.11.4
gunicorn==21.2.0